requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
scikit-learn>=0.24.0
scipy>=1.7.0
tqdm>=4.60.0
PyYAML>=5.4.0
matplotlib>=3.3.0
seaborn>=0.11.0
wandb>=0.12.0
Pillow>=8.0.0
